+++ tests/invariants/BasicInvariants.t.sol
@@ -29,7 +29,7 @@ contract BasicInvariants is BaseInvariants {
 
         currentTimestamp = block.timestamp;
 
-        for (uint i; i < NUM_LPS; i++) {
+        for (uint256 i; i < NUM_LPS; i++) {
             lps.push(makeAddr(string(abi.encode("lp", i))));
         }
